CRUD for asset

+++ src/main/java/com/example/table/repository/AssetRepository.java
@@ -3,5 +3,5 @@ package com.example.table.repository;
 import com.example.table.model.Asset;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AssetRepository extends JpaRepository<Asset, Integer> {
+public interface AssetRepository extends JpaRepository<Asset, Long> {
 }
